server/tcd_back/requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
django-otp==1.0.4
django-two-factor-auth==1.14
phonenumbers==8.12.34
qrcode[pil]==6.1
pyotp==2.6.0
channels==4.0.0